# apps/core/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoreX Service API",
    "description": "HTTP 状态码始终为 200；业务结果见响应体 body.code（200000=成功）。错误码规则见 /guide/error-codes.md。Exception.details 字段仅在开发环境返回。鉴权：JWT Bearer；Apifox 环境变量统一使用 {{token}}（来自 signin.data.token）。密码登录 / 注册 / 发 OTP 前须先 POST /api/v1/auth/captcha 获取 captchaKey（go-captcha 行为验证码）。",
    "contact": {
      "name": "CoreX Team",
      "email": "15638470820@163.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "development"
    },
    {
      "url": "https://api.example.com",
      "description": "production"
    }
  ],
  "paths": {
...
    "/api/v1/auth/captcha": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "获取行为验证码",
        "description": "公开接口。按客户端 IP 限流（`auth.captcha.ip_rate_limit` / 分钟）。\n\n后端代理 **go-captcha-service** `get-data`，返回滑块拼图数据，供前端 `go-captcha-react` 渲染。\n\n**联调流程（signin / signup / otp 前必做）：**\n1. 调用本接口（可选 body `kind`，默认 `auth.captcha.kind` 如 `slide-default`）\n2. 取 `data.captchaKey`、`masterImage`、`thumbImage`、`thumbX`/`thumbY` 等\n3. 用户完成交互后，在后续请求填写 `captchaKey` + `captchaValue`\n\n校验由侧车 `check-data` 完成；TTL 由侧车 Redis 缓存管理。",
        "operationId": "auth.captcha",
        "requestBody": {
          "description": "可选：指定题型 ID",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptchaP"
              },
              "example": {
                "kind": "slide-default"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "缓存异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/password": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "修改密码",
        "description": "需要 JWT。校验原密码后更新；成功后吊销当前 token，须重新登录。",
        "operationId": "auth.password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/auth/profile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "获取个人信息",
        "description": "需要 JWT 鉴权，返回当前登录用户的 profile（含 avatar 元数据）。",
        "operationId": "auth.toRead",
        "responses": {
          "200": {
            "description": "未登录（code=300001）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "更新个人信息",
        "description": "需要 JWT 鉴权。可更新 `email`、`phone`、`gender`、`birthday`、`avatar`（asset UUID）。\n所有字段可选；`avatar` 传 `null` 可清空头像。",
        "operationId": "auth.toUpdate",
        "requestBody": {
          "description": "可更新字段（均为可选）",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileP"
              },
              "example": {
                "avatar": "550e8400-e29b-41d4-a716-446655440000",
                "birthday": "1990-01-01",
                "email": "admin@example.com",
                "gender": "MALE",
                "phone": "13800138000"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "未登录或参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/auth/signout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "用户登出",
        "description": "需要 JWT。将当前 token 写入 Redis 黑名单直至过期，之后该 token 不可再用。",
        "operationId": "auth.signout",
        "responses": {
          "200": {
            "description": "未登录或缓存异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/tenants/{id}": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "summary": "租户详情",
        "operationId": "tenant.toReadById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Tenant"
        ],
        "summary": "更新租户",
        "operationId": "tenant.toUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "租户信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUpdateP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Tenant"
        ],
        "summary": "删除租户",
        "operationId": "tenant.toRemove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "租户 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "获取用户详情",
        "operationId": "user.toReadById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 UUID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3c430c21-0891-43e1-bcd2-1a22eb4a5389"
          }
        ],
        "responses": {
          "200": {
            "description": "用户不存在（code=500101）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "更新用户",
        "operationId": "user.toUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 UUID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3c430c21-0891-43e1-bcd2-1a22eb4a5389"
          }
        ],
        "requestBody": {
          "description": "可更新字段（均为可选）",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateP"
              },
              "example": {
                "age": 30,
                "birthday": "1990-01-01",
                "email": "alice@example.com",
                "gender": "FEMALE",
                "phone": "13800138000",
                "role": "USER",
                "status": "ACTIVE",
                "username": "alice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "用户不存在或参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "删除用户",
        "operationId": "user.toRemove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "用户 UUID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "3c430c21-0891-43e1-bcd2-1a22eb4a5389"
          }
        ],
        "responses": {
          "200": {
            "description": "用户不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CaptchaP": {
        "type": "object",
        "properties": {
          "kind": {
            "type": [
              "string",
              "null"
            ],
            "description": "题型 ID，默认 `auth.captcha.kind`",
            "example": "slide-default"
          }
        }
      },
...
      "Exception": {
        "type": "object",
        "description": "异常响应结构",
        "required": [
          "code",
          "success",
          "msg",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "业务错误码 (非0表示各种业务错误)"
          },
          "data": {
            "type": [
              "object",
              "null"
            ]
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "description": "详细错误信息（仅非生产环境返回）"
          },
          "msg": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PasswordP": {
        "type": "object",
        "required": [
          "oldPassword",
          "newPassword"
        ],
        "properties": {
          "newPassword": {
            "type": "string",
            "example": "654321"
          },
          "oldPassword": {
            "type": "string",
            "example": "123456"
          }
        },
        "example": {
          "newPassword": "654321",
          "oldPassword": "123456"
        }
      },
...
      "ProfileP": {
        "type": "object",
        "properties": {
          "avatar": {
            "type": [
              "string",
              "null"
            ],
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "birthday": {
            "type": [
              "string",
              "null"
            ],
            "example": "1990-01-01"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "example": "admin@example.com"
          },
          "gender": {
            "type": [
              "string",
              "null"
            ],
            "example": "MALE"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "example": "13800138000"
          }
        },
        "example": {
          "avatar": "550e8400-e29b-41d4-a716-446655440000",
          "birthday": "1990-01-01",
          "email": "admin@example.com",
          "gender": "MALE",
          "phone": "13800138000"
        }
      },
...
      "TenantUpdateP": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": [
              "string",
              "null"
            ],
            "description": "PERSONAL / TEAM"
          }
        }
      },
...
      "UpdateP": {
        "type": "object",
        "description": "后台管理员更新用户（账号/密码/资料运维）\n用户自助修改个人信息请使用 PUT /api/v1/auth/profile",
        "properties": {
          "age": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "example": 30,
            "minimum": 0
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ],
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "birthday": {
            "type": [
              "string",
              "null"
            ],
            "example": "1990-01-01"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "example": "alice@example.com"
          },
          "gender": {
            "type": [
              "string",
              "null"
            ],
            "example": "FEMALE"
          },
          "password": {
            "type": [
              "string",
              "null"
            ],
            "example": "123456"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "example": "13800138000"
          },
          "role": {
            "type": [
              "string",
              "null"
            ],
            "example": "USER"
          },
          "status": {
            "type": [
              "string",
              "null"
            ],
            "example": "ACTIVE"
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "example": "alice"
          }
        },
        "example": {
          "age": 30,
          "birthday": "1990-01-01",
          "email": "alice@example.com",
          "gender": "FEMALE",
          "phone": "13800138000",
          "role": "USER",
          "status": "ACTIVE",
          "username": "alice"
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录 `POST /api/v1/auth/signin` 返回的 `data.token`。\n\n请求头：`Authorization: Bearer <token>`。\n\n**Apifox**：鉴权组件 Token 请填 `{{token}}`（勿用 `{{bearerToken}}`）；环境变量名统一为 `token`，可在登录接口后置操作写入。",
        "x-default": "{{token}}",
        "x-apifox-default": "{{token}}"
      }
    }
  },
...
}